Fixed a bug where looping through the CallbackList components would crash the engine

+++ AstralEngine/src/AstralEngine/Renderer/VertexBuffer.h
@@ -81,7 +81,7 @@ namespace AstralEngine
 
 		unsigned int GetStride() const { return m_stride; }
 		
-		inline int GetCount() const { return m_elements.GetCount(); }
+		inline size_t GetCount() const { return m_elements.GetCount(); }
 
 		const LayoutElement& operator[](unsigned int index) const
 		{
